requirements.txt
# Core dependencies
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database and caching
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0

# Media handling
requests>=2.32.2
Pillow==10.2.0
python-magic==0.4.27
yt-dlp==2023.11.16

# Performance optimization
uvicorn==0.27.1
gunicorn==21.2.0
whitenoise==6.6.0

# Security
cryptography==42.0.2
django-ratelimit==4.1.0

# Utils
python-dotenv==1.0.1
aiohttp==3.9.3